[aptos-rosetta] Share block lookups between account and block APIs
